extract modals to avoid duplicate rendering and overhead
